Improved exercise POST

+++ src/main/java/com/springapp/controller/exercise/ExerciseResourceBean.java
@@ -4,8 +4,10 @@ import com.springapp.exception.ExerciseNotFoundException;
 import com.springapp.jpa.model.Exercise;
 import com.springapp.jpa.repository.ExerciseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,8 +15,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.validation.Valid;
+import java.net.URI;
 import java.util.Optional;
 
 /**
@@ -40,11 +44,23 @@ public class ExerciseResourceBean implements ExerciseResource {
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    @ResponseStatus(HttpStatus.CREATED)
-    public Exercise processRegistration(@Valid
-                                        @RequestBody
-                                        final Exercise exercise) {
-        return exerciseRepository.save(exercise);
+    public ResponseEntity<Exercise> processRegistration(@Valid
+                                                        @RequestBody
+                                                        final Exercise exercise) {
+        final Exercise savedExercise = exerciseRepository.save(exercise);
+
+        final HttpHeaders headers = new HttpHeaders();
+        final URI locationUri = ServletUriComponentsBuilder
+                .fromCurrentServletMapping().path("/exercise/")
+                .path(String.valueOf(savedExercise.getId()))
+                .build()
+                .toUri();
+
+        headers.setLocation(locationUri);
+
+        final ResponseEntity<Exercise> responseEntity =
+                new ResponseEntity<>(savedExercise, headers, HttpStatus.CREATED);
+        return responseEntity;
     }
 
     /**
@@ -64,7 +80,7 @@ public class ExerciseResourceBean implements ExerciseResource {
             return result.get();
         }
 
-        return null;
+        throw new ExerciseNotFoundException(id);
     }
 
     /**
@@ -77,6 +93,12 @@ public class ExerciseResourceBean implements ExerciseResource {
     @ResponseStatus(HttpStatus.OK)
     public void deleteExerciseById(@PathVariable
                                    final Long id) {
+        final Optional<Exercise> result = exerciseRepository.findOne(id);
+
+        if (!result.isPresent()) {
+            throw new ExerciseNotFoundException(id);
+        }
+
         exerciseRepository.delete(id);
     }
